feat: apply first draft

+++ CPP/Clipper2Lib/include/clipper2/clipper.engine.h
@@ -176,7 +176,7 @@ namespace Clipper2Lib {
 
         //  we want callback that returns group_delta depending on path and point index
         //  and we want to be able to pass in a group_delta value
-        typedef std::function<int64_t(const Path64& path, size_t pt_idx)> DeltaCallback64;
+        typedef std::function<double(const Path64& path, size_t pt_idx)> DeltaCallback64;
         typedef std::function<double(const PathD& path, size_t pt_idx)> DeltaCallbackD;
 
 	typedef std::vector<HorzSegment> HorzSegmentList;
